+++ packages/camera/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.5.1
+
+* Can now be compiled with earlier Android sdks when
+`<uses-sdk tools:overrideLibrary="io.flutter.plugins.camera"/>` has been added to the project
+`AndroidManifest.xml`.
+
 ## 0.5.0
 
 * **Breaking Change** This plugin no longer handles closing and opening the camera on Android